price_comparison_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
pyjwt==2.8.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
psycopg2-binary==2.9.9
oracledb==2.0.0